Updated base.py - Removed extra argument from get_nearest_sampling_interval() - Allowed ctypes to auto-select n_captures in set_no_of_captures()

+++ pypicosdk/base.py
@@ -651,7 +651,7 @@ class PicoScopeBase:
         self._call_attr_function(
             "SetNoOfCaptures",
             self.handle,
-            ctypes.c_uint64(n_captures),
+            n_captures,
         )
 
     def get_no_of_captures(self) -> int:
